requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2

# NLP and text processing
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1

# OCR for menu images
pytesseract==0.3.10
easyocr==1.7.1
Pillow==10.1.0

# Data handling
pandas==2.1.3
numpy==1.26.2

# LLM APIs (choose based on preference)
openai==1.3.7
google-generativeai==0.3.1
anthropic==0.7.7

# Web scraping utilities
lxml==4.9.3
cloudscraper==1.2.71

# Testing and validation
pytest==7.4.3
